Remove the warning x.6 BADGE=linux BUILD_TYPE=cross C_COMPILER=x86_64-w64-mingw32-gcc CXX_COMPILER=x86_64-w64-mingw32-g++

+++ src/align_tools.cpp
@@ -91,7 +91,7 @@ chunk_t *scan_ib_line(chunk_t *start, bool first_pass)
             if (cpd.al_cnt == AL_SIZE)
             {
                fprintf(stderr, "Number of 'entry' to be aligned is too big for the current value %d,\n", AL_SIZE);
-               fprintf(stderr, "at line %zu, column %zu.\n", pc->orig_line, pc->orig_col);
+               fprintf(stderr, "at line %d, column %d.\n", (int)pc->orig_line, (int)pc->orig_col);
                fprintf(stderr, "Please make a report.\n");
                log_flush(true);
                exit(EX_SOFTWARE);
